realtime-reseach-copilot/requirements.txt
# Core web framework
fastapi
uvicorn[standard]

# Frontend
streamlit

# Local LLM inference (no OpenAI)
ollama

# Web search (optional - only used if TAVILY_API_KEY is set)
tavily-python

# Free encyclopedia search
wikipedia

# Data validation and settings
pydantic>=2.0
pydantic-settings
python-dotenv

# HTTP client (used for health checks and frontend→backend calls)
requests
httpx

# Vector search for private knowledge base
faiss-cpu
numpy

# Testing
pytest
pytest-asyncio
